[Update] Move src/lib/*.js to src/utils/
