+++ src/components/Daily.js
@@ -1,7 +1,7 @@
 import { Link } from 'react-router'
 import React, { Component } from 'react'
 import { LINK_PREFIX, INTERACTIVE_ARTICLE_STYLE } from '../constants/index'
-import { date2yyyymmdd } from '../lib/date-transformer'
+import { date2yyyymmdd } from '../utils/index'
 import { getArticleImageSrc } from '../utils/index'
 
 // lodash
