+++ src/components/TopNews.js
@@ -1,6 +1,6 @@
 'use strict'
 import { LINK_PREFIX, INTERACTIVE_ARTICLE_STYLE } from '../constants/index'
-import { date2yyyymmdd } from '../lib/date-transformer'
+import { date2yyyymmdd } from '../utils/index'
 import { getArticleImageSrc, getArticleImageSrcSet } from '../utils/index'
 import { Link } from 'react-router'
 import Hexagon from './Hexagon'
